save user while creating job

+++ job/views.py
@@ -1,11 +1,11 @@
 from django.db.models import Count, Avg, Min, Max
 from django.shortcuts import render
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
-
+from rest_framework.permissions import IsAuthenticated
 from .serializers import JobSerializer
 from .models import Job
 from .filters import JobFilter
@@ -41,7 +41,9 @@ def getJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
 
     job = Job.objects.create(**data)
@@ -51,8 +53,12 @@ def newJob(request):
     return Response(serializer.data)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You cant update this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.title           = request.data['title']
     job.description     = request.data['description']
@@ -73,8 +79,12 @@ def updateJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You cant delete this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.delete()
